Update to move SimpleItem to package level
